Create color generator
In this code, we have an array called colors that stores the colors of the rainbow. The getRandomColor function generates a random index within the range of the colors array length and returns the corresponding color.

You can run this code in a JavaScript environment (e.g., browser console or Node.js) to see the randomly selected color printed in the console.